RUMM-163 Limit the size of logs folder

+++ Tests/DatadogTests/Mocks/FoundationMocks.swift
@@ -36,8 +36,8 @@ extension Data {
         return Data(repeating: byte, count: count)
     }
 
-    static func mock(ofSize size: Int) -> Data {
-        return mockRepeating(byte: 0x41, times: size)
+    static func mock(ofSize size: UInt64) -> Data {
+        return mockRepeating(byte: 0x41, times: Int(size))
     }
 }
